Fix compilation crash in older versions of Clang caused by extreme comma-initializer use.
The `cube_mat` matrix in `LazyCubicInterpolator` has 4096 entries in the 3D case.
Entering all of these values using Eigen's comma-initializer syntax
initiates a large number of template instantiations that crashes
Clang version 19 and older (as well as Apple Clang 16 and older)
and triggers a warning in the latest versions.

This commit fixes the crash by switching to storing the values in a
compile-time-constant lookup table and mapping the array as an Eigen
matrix. Since the mapping can be done on the fly, we removed
the `cubic_mat` member variable.

While at it, we also changed the `x` vector to a fixed-size vector type
to avoid a memory allocation. There are many other usages of
`MatrixXd` in this class that could be static, fixed-size types--not to
mention significant sparsity in `cubic_mat_data` that could be
exploited--but we'll leave it at here for now.